Remove support for module linking
This gets tests to pass now that bytecodealliance/wasmtime#3958 has
landed.
